reorganize Site class to be in a single file
