+++ Gemfile
@@ -4,5 +4,7 @@ gem 'pry'
 gem 'chunky_png'
 gem 'selenium-webdriver'
 gem 'httparty'
+gem 'phantomjs'
+
 # required to vendor because this gem is not publicly accessible and should not be for the foreseable future
 gem 'core-auth-ruby', '~> 0.2.0', path: 'vendor/gems/core-auth-ruby-0.2.0'
