+++ snapper-caller.rb
@@ -4,6 +4,7 @@ $:.unshift(File.expand_path(File.dirname(__FILE__), 'lib'))
 
 require 'optparse'
 require 'snapper'
+require 'site'
 
 options = {}
